control-atlas-saas/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.30.0
plotly==5.18.0
pandas==2.1.4
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.1.7
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
fpdf2==2.7.7
celery==5.3.6
redis==5.0.1
biopython==1.81
numpy>=1.24.0
scipy>=1.10.0
gemmi==0.6.4
openmm>=8.1.1
git+https://github.com/openmm/pdbfixer.git
